+++ lib/Controller/PageController.php
@@ -25,34 +25,43 @@ namespace OCA\Contacts\Controller;
 
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http\TemplateResponse;
-// use OCP\IInitialStateService;
+use OCP\IInitialStateService;
 use OCP\IConfig;
-use OCP\L10N\IFactory;
 use OCP\IRequest;
+use OCP\L10N\IFactory;
+use OCP\Util;
+use OCA\Contacts\Controller\SocialApiController;
 
 class PageController extends Controller {
 
 	protected $appName;
 
-	// /** @var IInitialStateService */
-	// private $initialStateService;
-
-	/** @var IFactory */
-	private $languageFactory;
 	/** @var IConfig */
 	private  $config;
 
-	public function __construct(string $AppName,
+	/** @var IInitialStateService */
+	private $initialStateService;
+
+	/** @var IFactory */
+	private $languageFactory;
+
+	/** @var SocialApiController */
+	private $socialApi;
+
+	public function __construct(string $appName,
 								IRequest $request,
 								IConfig $config,
-								// IInitialStateService $initialStateService,
-								IFactory $languageFactory) {
-		parent::__construct($AppName, $request);
+								IInitialStateService $initialStateService,
+								IFactory $languageFactory,
+								SocialApiController $socialApi) {
+		parent::__construct($appName, $request);
 
-		$this->appName = $AppName;
-		// $this->initialStateService = $initialStateService;
-		$this->languageFactory = $languageFactory;
+		$this->appName = $appName;
 		$this->config = $config;
+		$this->initialStateService = $initialStateService;
+		$this->languageFactory = $languageFactory;
+
+		$this->socialApi = $socialApi;
 	}
 
 	/**
@@ -64,12 +73,15 @@ class PageController extends Controller {
 	public function index(): TemplateResponse {
 		$locales = $this->languageFactory->findAvailableLocales();
 		$defaultProfile = $this->config->getAppValue($this->appName, 'defaultProfile', 'HOME');
-		// TODO: use initialStateService once min-version is 16!
-		// $this->initialStateService->provideInitialState($this->appName, 'locales', $locales);
-		// $this->initialStateService->provideInitialState($this->appName, 'defaultProfile', $defaultProfile);
-		return new TemplateResponse(
-			'contacts',
-			'main',
-			['locales' => json_encode($locales), 'defaultProfile'=> json_encode($defaultProfile)]); // templates/main.php
+		$supportedNetworks = $this->socialApi->getSupportedNetworks();
+
+		$this->initialStateService->provideInitialState($this->appName, 'locales', $locales);
+		$this->initialStateService->provideInitialState($this->appName, 'defaultProfile', $defaultProfile);
+		$this->initialStateService->provideInitialState($this->appName, 'supportedNetworks', $supportedNetworks);
+
+		Util::addScript($this->appName, 'contacts');
+		Util::addStyle($this->appName, 'contacts');
+
+		return new TemplateResponse($this->appName, 'main');
 	}
 }