+++ src/components/ContactDetails.vue
@@ -39,7 +39,9 @@
 			<!-- contact header -->
 			<header :style="{ 'backgroundColor': colorAvatar }">
 				<!-- avatar and upload photo -->
-				<ContactAvatar :contact="contact" />
+				<ContactAvatar
+					:contact="contact"
+					@updateLocalContact="updateLocalContact" />
 				<!-- QUESTION: is it better to pass contact as a prop or get it from the store inside
 				contact-avatar ?  :avatar="contact.photo"-->
 
@@ -94,7 +96,7 @@
 						:class="{'icon-loading-small': loadingUpdate,
 							[`${warning.icon}`]: warning}"
 						class="header-icon"
-						href="#" />
+						@click="onWarningClick" />
 
 					<!-- conflict message -->
 					<div v-if="conflict"
@@ -117,16 +119,22 @@
 						@click="updateContact" />
 
 					<!-- menu actions -->
-					<Actions class="header-menu" menu-align="right">
+					<Actions ref="actions"
+						class="header-menu"
+						menu-align="right"
+						:open.sync="openedMenu">
 						<ActionLink :href="contact.url"
 							:download="`${contact.displayName}.vcf`"
 							icon="icon-download">
 							{{ t('contacts', 'Download') }}
 						</ActionLink>
-						<!-- FIXME: show this menu item only if respective field available -->
-						<!-- TODO: find a better icon -->
-						<ActionButton icon="icon-download" @click="downloadSocialPic">
-							{{ t('contacts', 'Get Profile Picture from Facebook') }}
+						<!-- user can clone if there is at least one option available -->
+						<ActionButton v-if="isReadOnly && addressbooksOptions.length > 0"
+							ref="cloneAction"
+							:close-after-click="true"
+							icon="icon-clone"
+							@click="cloneContact">
+							{{ t('contacts', 'Clone contact') }}
 						</ActionButton>
 						<ActionButton icon="icon-qrcode" @click="showQRcode">
 							{{ t('contacts', 'Generate QR Code') }}
@@ -140,9 +148,34 @@
 				<!-- qrcode -->
 				<Modal v-if="qrcode"
 					id="qrcode-modal"
+					:clear-view-delay="-1"
 					:title="contact.displayName"
 					@close="closeQrModal">
-					<img :src="`data:image/svg+xml;base64,${qrcode}`" class="qrcode" width="400">
+					<img :src="`data:image/svg+xml;base64,${qrcode}`"
+						:alt="t('contacts', 'Contact vCard as qrcode')"
+						class="qrcode"
+						width="400">
+				</Modal>
+
+				<!-- pick addressbook when cloning contact -->
+				<Modal v-if="showPickAddressbookModal"
+					id="pick-addressbook-modal"
+					:clear-view-delay="-1"
+					:title="t('contacts', 'Pick an address book')"
+					@close="closePickAddressbookModal">
+					<Multiselect ref="pickAddressbook"
+						v-model="pickedAddressbook"
+						:allow-empty="false"
+						:options="addressbooksOptions"
+						:placeholder="t('contacts', 'Select address book')"
+						track-by="id"
+						label="name" />
+					<button @click="closePickAddressbookModal">
+						{{ t('contacts', 'Cancel') }}
+					</button>
+					<button class="primary" @click="cloneContact">
+						{{ t('contacts', 'Clone contact') }}
+					</button>
 				</Modal>
 			</header>
 
@@ -198,7 +231,12 @@ import debounce from 'debounce'
 import PQueue from 'p-queue'
 import qr from 'qr-image'
 import { stringify } from 'ical.js'
-import { ActionLink, ActionButton } from '@nextcloud/vue'
+import Actions from '@nextcloud/vue/dist/Components/Actions'
+import ActionLink from '@nextcloud/vue/dist/Components/ActionLink'
+import ActionButton from '@nextcloud/vue/dist/Components/ActionButton'
+import Multiselect from '@nextcloud/vue/dist/Components/Multiselect'
+import Modal from '@nextcloud/vue/dist/Components/Modal'
+import { showError } from '@nextcloud/dialogs'
 
 import rfcProps from '../models/rfcProps'
 import validate from '../services/validate'
@@ -216,6 +254,7 @@ export default {
 	name: 'ContactDetails',
 
 	components: {
+		Actions,
 		ActionButton,
 		ActionLink,
 		AddNewProp,
@@ -224,6 +263,8 @@ export default {
 		PropertyGroups,
 		PropertyRev,
 		PropertySelect,
+		Modal,
+		Multiselect,
 	},
 
 	props: {
@@ -252,6 +293,8 @@ export default {
 			loadingUpdate: false,
 			openedMenu: false,
 			qrcode: '',
+			showPickAddressbookModal: false,
+			pickedAddressbook: null,
 		}
 	},
 
@@ -331,7 +374,7 @@ export default {
 		 */
 		addressbookModel() {
 			return {
-				readableName: t('contacts', 'Addressbook'),
+				readableName: t('contacts', 'Address book'),
 				icon: 'icon-address-book',
 				options: this.addressbooksOptions,
 			}
@@ -382,6 +425,7 @@ export default {
 
 		/**
 		 * Store getters filtered and mapped to usable object
+		 * This is the list of addressbooks that are available to write
 		 *
 		 * @returns {Array}
 		 */
@@ -479,19 +523,6 @@ export default {
 		},
 
 		/**
-		 * Import facebook profile pic
-		 */
-		downloadSocialPic() {
-			const jCal = this.contact.jCal.slice(0)
-			const facebookid = jCal[1].filter(props => props[0] === 'x-socialprofile')
-			if (facebookid.length) {
-				console.debug('https://graph.facebook.com/' + facebookid[0][3] + '/picture?type=large')
-				// TODO: upload picture
-			}
-			// TODO: error handling
-		},
-
-		/**
 		 * Select the text in the input if it is still set to 'new Contact'
 		 */
 		selectInput() {
@@ -522,11 +553,11 @@ export default {
 						await this.updateLocalContact(contact)
 					} catch (error) {
 						if (error.name === 'ParserError') {
-							OC.Notification.showTemporary(t('contacts', 'Syntax error. Cannot open the contact.'))
+							showError(t('contacts', 'Syntax error. Cannot open the contact.'))
 						} else if (error.status === 404) {
-							OC.Notification.showTemporary(t('contacts', `The contact doesn't exists anymore on the server.`))
+							showError(t('contacts', `The contact doesn't exists anymore on the server.`))
 						} else {
-							OC.Notification.showTemporary(t('contacts', `Unable to retrieve the contact from the server, please check your network connection.`))
+							showError(t('contacts', `Unable to retrieve the contact from the server, please check your network connection.`))
 						}
 						console.error(error)
 						// trigger a local deletion from the store only
@@ -574,6 +605,41 @@ export default {
 					})
 				} catch (error) {
 					console.error(error)
+					showError(t('contacts', 'An error occurred while trying to move the contact'))
+				} finally {
+					this.loadingUpdate = false
+				}
+			}
+		},
+
+		/**
+		 * Copy contact to the specified addressbook
+		 *
+		 * @param {string} addressbookId the desired addressbook ID
+		 */
+		async copyContactToAddressbook(addressbookId) {
+			const addressbook = this.addressbooks.find(search => search.id === addressbookId)
+			this.loadingUpdate = true
+			if (addressbook) {
+				try {
+					const contact = await this.$store.dispatch('copyContactToAddressbook', {
+						// we need to use the store contact, not the local contact
+						// using this.contact and not this.localContact
+						contact: this.contact,
+						addressbook,
+					})
+					// select the contact again
+					this.$router.push({
+						name: 'contact',
+						params: {
+							selectedGroup: this.$route.params.selectedGroup,
+							selectedContact: contact.key,
+						},
+					})
+				} catch (error) {
+					console.error(error)
+					window.temp1 = error
+					showError(t('contacts', 'An error occurred while trying to copy the contact'))
 				} finally {
 					this.loadingUpdate = false
 				}
@@ -616,6 +682,42 @@ export default {
 			if (e.keyCode === 83 && (navigator.platform.match('Mac') ? e.metaKey : e.ctrlKey)) {
 				e.preventDefault()
 				this.debounceUpdateContact()
+			}
+		},
+
+		/**
+		 * Clone the current contact to another addressbook
+		 */
+		async cloneContact() {
+			// only one addressbook, let's clone it there
+			if (this.pickedAddressbook && this.addressbooks.find(addressbook => addressbook.id === this.pickedAddressbook.id)) {
+				console.debug('Cloning contact to', this.pickedAddressbook.name)
+				await this.copyContactToAddressbook(this.pickedAddressbook.id)
+				this.closePickAddressbookModal()
+			} else if (this.addressbooksOptions.length === 1) {
+				console.debug('Cloning contact to', this.addressbooksOptions[0].name)
+				await this.copyContactToAddressbook(this.addressbooksOptions[0].id)
+			} else {
+				this.showPickAddressbookModal = true
+			}
+		},
+
+		closePickAddressbookModal() {
+			this.showPickAddressbookModal = false
+			this.pickedAddressbook = null
+		},
+
+		/**
+		 * The user clicked the warning icon
+		 */
+		onWarningClick() {
+			// if the user clicked the readonly icon, let's focus the clone button
+			if (this.isReadOnly && this.addressbooksOptions.length > 0) {
+				this.openedMenu = true
+				this.$nextTick(() => {
+					// focus the clone button
+					this.$refs.actions.onMouseFocusAction({ target: this.$refs.cloneAction.$el })
+				})
 			}
 		},
 	},