+++ src/components/Settings/SettingsAddressbook.vue
@@ -93,18 +93,23 @@
 </template>
 
 <script>
-import { ActionLink, ActionButton, ActionInput, ActionCheckbox } from '@nextcloud/vue'
+import Actions from '@nextcloud/vue/dist/Components/Actions'
+import ActionLink from '@nextcloud/vue/dist/Components/ActionLink'
+import ActionButton from '@nextcloud/vue/dist/Components/ActionButton'
+import ActionInput from '@nextcloud/vue/dist/Components/ActionInput'
+import ActionCheckbox from '@nextcloud/vue/dist/Components/ActionCheckbox'
 import ShareAddressBook from './SettingsAddressbookShare'
 
 export default {
 	name: 'SettingsAddressbook',
 
 	components: {
-		ShareAddressBook,
-		ActionLink,
 		ActionButton,
-		ActionInput,
 		ActionCheckbox,
+		ActionInput,
+		ActionLink,
+		Actions,
+		ShareAddressBook,
 	},
 
 	props: {
@@ -190,7 +195,7 @@ export default {
 			} catch (err) {
 				// error handling
 				console.error(err)
-				OC.Notification.showTemporary(t('contacts', 'Toggling of addressbook was not successful'))
+				OC.Notification.showTemporary(t('contacts', 'Toggling of address book was not successful'))
 			} finally {
 				// stop loading status regardless of outcome
 				this.toggleEnabledLoading = false
@@ -199,7 +204,7 @@ export default {
 
 		confirmDeletion() {
 			OC.dialogs.confirm(
-				t('contacts', 'This will delete the addressbook and every contacts within it'),
+				t('contacts', 'This will delete the address book and every contacts within it'),
 				t('contacts', 'Delete {addressbook} ?', { addressbook: this.addressbook.displayName }),
 				this.deleteAddressbook,
 				true
@@ -215,7 +220,7 @@ export default {
 				} catch (err) {
 					// error handling
 					console.error(err)
-					OC.Notification.showTemporary(t('contacts', 'Deletion of addressbook was not successful.'))
+					OC.Notification.showTemporary(t('contacts', 'Deletion of address book was not successful.'))
 				} finally {
 					// stop loading status regardless of outcome
 					this.deleteAddressbookLoading = false
@@ -236,7 +241,7 @@ export default {
 			} catch (err) {
 				// error handling
 				console.error(err)
-				OC.Notification.showTemporary(t('contacts', 'Renaming of addressbook was not successful.'))
+				OC.Notification.showTemporary(t('contacts', 'Renaming of address book was not successful.'))
 			} finally {
 				this.editingName = false
 				// stop loading status regardless of outcome
@@ -255,11 +260,11 @@ export default {
 				this.copySuccess = true
 				this.copied = true
 				// Notify addressbook was copied
-				OC.Notification.showTemporary(t('contacts', 'Addressbook copied to clipboard'))
+				OC.Notification.showTemporary(t('contacts', 'Address book copied to clipboard'))
 			} catch (error) {
 				this.copySuccess = false
 				this.copied = true
-				OC.Notification.showTemporary(t('contacts', 'Addressbook was not copied to clipboard.'))
+				OC.Notification.showTemporary(t('contacts', 'Address book was not copied to clipboard.'))
 			} finally {
 				this.copyLoading = false
 				setTimeout(() => {